+++ tests/negative/phase_a_malformed/malformed_macro.cpp
@@ -1,0 +1,10 @@
+#define AZTECA_BROKEN_EXPR(value) ((value) +
+
+class BrokenMacroHost
+{
+   public:
+	int run(int value)
+	{
+		return AZTECA_BROKEN_EXPR(value);
+	}
+};